We now _always_ log messages, we still need to check config if we want to log as well as save messages when logfile is not yet open, (limit the number kept in memory).

+++ myodd/log/log.cpp
@@ -172,7 +172,7 @@ namespace myodd{ namespace log{
     while( iSend >= 0 )
     {
       _LogMessage& lm = m_logMessages[ iSend-- ];
-      fnNotif( lm.GetType(), lm.GetMessage(), (void*)lParam );
+      fnNotif((unsigned int)lm.GetType(), lm.GetMessage(), (void*)lParam );
     }
 
     // add it to the list
